Break out and fix stack-value and formatting
