+++ ts/flexbuffers/flexbuffers-util.ts
@@ -1,9 +1,9 @@
 export function fromUTF8Array(data: BufferSource): string {
-    const decoder = new TextDecoder();
-    return decoder.decode(data);
+  const decoder = new TextDecoder();
+  return decoder.decode(data);
 }
 
 export function toUTF8Array(str: string) : Uint8Array {
-    const encoder = new TextEncoder();
-    return encoder.encode(str);
+  const encoder = new TextEncoder();
+  return encoder.encode(str);
 }
